Changed SDACD to return the final beta also
I changed SDACD tp return the final beta in addition to the B and Q matrices since the fianl beta was not being saved in the B matrix. I have been trying to run the test file with the output of SDACD for the training set from OOdata, but I keep having errors with the matrix dimensions.