Adds functions testAP and test_prox_EN
I added functions testAP and test_prox_EN to run the proximal gradient algorithm with the form of the problem from the paper to match the form for CD. The betaCoordDesc seems to work well with the 2 class case. For the 4 class case, I keep getting different stats when I run it. Sometimes it will have a good error, but many times it does not.  I also added 2 text files with the output from the command line to show how the error was changing. One file was before I started a diary.